Feature/cldn 1761, upgrade to 400 from upstream, refactoring of iceberg plugin for hadoop catalog , running pipeline into container

+++ core/trino-spi/src/main/java/io/trino/spi/type/SmallintType.java
@@ -138,7 +138,7 @@ public final class SmallintType
     @Override
     public long getLong(Block block, int position)
     {
-        return (long) block.getShort(position, 0);
+        return block.getShort(position, 0);
     }
 
     @Override
